refactor: move tests next to related code
