+++ src/print/lit.rs
@@ -22,3 +22,74 @@ impl<'a, 'b> Printer<'a, 'b> {
         self.print_attr_comment(html_lit.span().end());
     }
 }
+
+#[cfg(test)]
+mod test {
+    use crate::testing::*;
+
+    test_default!(
+        lit,
+        r#"
+        html!{ "Hello world!" }
+        "#,
+        r#"
+        html! {
+            "Hello world!"
+        }
+        "#
+    );
+
+    // NOTE: multiline string formatting is left to the users
+    test_default!(
+        whitespace_in_multi_line_strings_edge_case,
+        r##"
+        html! {
+        p {
+            (PreEscaped(r#"
+            Multiline
+
+            String
+            "#))
+        }
+        }
+        "##,
+        r##"
+        html! {
+            p {
+                ({
+                    PreEscaped(
+                        r#"
+            Multiline
+
+            String
+            "#,
+                    )
+                })
+            }
+        }
+        "##
+    );
+
+    // NOTE: multiline string formatting is left to the users
+    test_default!(
+        correct_multiline_string_indent_in_splices,
+        r##"
+        html! {
+            (r#"
+            Multiline
+            String
+            "#)
+        }
+        "##,
+        r##"
+        html! {
+            ({
+                r#"
+            Multiline
+            String
+            "#
+            })
+        }
+        "##
+    );
+}
